功能添加：添加用户登录功能，和根据token查询用户信息功能

+++ bv-mall-parent/bv-mall-api-entity/bv-mall-api-member-entity/src/main/java/com/bv/member/dto/input/UserInDTO.java
@@ -22,7 +22,7 @@ public class UserInDTO implements Serializable {
      * 主键id
      */
     @ApiModelProperty(value = "主键id")
-    private Integer idMallUser;
+    private Long idMallUser;
 
     /**
      * 手机号
